Move medium to cbmodel
