+++ src/framed/core/warnings.py
@@ -1,0 +1,2 @@
+class OptimizationWarning(UserWarning):
+    pass
