+++ src/framed/io/sbml.py
@@ -15,6 +15,8 @@ from sympy import to_dnf, Or, And
 from sympy.logic.boolalg import is_dnf
 from libsbml import SBMLReader, SBMLWriter, SBMLDocument, XMLNode, AssignmentRule, parseL3FormulaWithModel, FbcExtension
 
+import warnings
+
 DEFAULT_SBML_LEVEL = 3
 DEFAULT_SBML_VERSION = 1
 
@@ -183,8 +185,7 @@ def _load_cbmodel(sbml_model, flavor):
         _load_fbc2_objective(sbml_model, model)
         _load_fbc2_gpr(sbml_model, model)
     else:
-        print 'unsupported SBML flavor:', flavor
-        print 'supported flavors:', COBRA_MODEL, FBC2_MODEL
+        raise TypeError("Unsupported SBML flavor '{}' (supported: {}".format(flavor, ",".join([COBRA_MODEL, FBC2_MODEL])))
 
     return model
 
@@ -324,13 +325,13 @@ def _load_fbc2_gpr(sbml_model, model):
         fbcrxn = reaction.getPlugin('fbc')
         gpr_assoc = fbcrxn.getGeneProductAssociation()
         if gpr_assoc:
-            gpr = _parse_fbc_association(gpr_assoc.getAssociation())
+            gpr = _parse_fbc_association(gpr_assoc.getAssociation(), reaction.id)
             model.set_gpr_association(reaction.getId(), gpr)
         else:
             model.set_gpr_association(reaction.getId(), None)
 
 
-def _parse_fbc_association(gpr_assoc):
+def _parse_fbc_association(gpr_assoc, reaction_id):
 
     gpr = GPRAssociation()
 
@@ -342,11 +343,13 @@ def _parse_fbc_association(gpr_assoc):
                     if subitem.isGeneProductRef():
                         protein.genes.append(subitem.getGeneProduct())
                     else:
-                        print 'Gene association is not DNF'
+                        w = "Gene association for reaction '{}' is not DNF".format(reaction_id)
+                        warnings.warn(w, SyntaxWarning)
             elif item.isGeneProductRef:
                 protein.genes.append(item.getGeneProduct())
             else:
-                print 'Gene association is not DNF'
+                w = "Gene association for reaction '{}' is not DNF".format(reaction_id)
+                warnings.warn(w, SyntaxWarning)
             gpr.proteins.append(protein)
 
     elif gpr_assoc.isFbcAnd():
@@ -355,14 +358,17 @@ def _parse_fbc_association(gpr_assoc):
             if item.isGeneProductRef():
                 protein.genes.append(item.getGeneProduct())
             else:
-                print 'Gene association is not DNF'
+                w = "Gene association for reaction '{}' is not DNF".format(reaction_id)
+                warnings.warn(w, SyntaxWarning)
+
         gpr.proteins = [protein]
     elif gpr_assoc.isGeneProductRef():
         protein = Protein()
         protein.genes = [gpr_assoc.getGeneProduct()]
         gpr.proteins = [protein]
     else:
-        print 'Gene association is not DNF'
+        w = "Gene association for reaction '{}' is not DNF".format(reaction_id)
+        warnings.warn(w, SyntaxWarning)
 
     return gpr
