+++ src/framed/model/cbmodel.py
@@ -1,7 +1,14 @@
-from collections import OrderedDict
+from collections import OrderedDict, MutableMapping
+from copy import copy, deepcopy
 
 from .model import Model, Metabolite, Reaction, AttrOrderedDict
 from .parser import ReactionParser
+
+from collections import OrderedDict, MutableMapping
+from copy import deepcopy
+import os, re, errno
+
+import warnings
 
 
 class Gene:
@@ -156,11 +163,24 @@ class CBModel(Model):
         """
         Model.__init__(self, model_id)
         self.genes = AttrOrderedDict()
-        self.biomass_reaction = None
+        self.__biomass_reaction = None
+        self.__biomass_reaction_detected = False
 
     def _clear_temp(self):
         Model._clear_temp(self)
         self._g_r_lookup = None
+
+    @property
+    def biomass_reaction(self):
+        if not self.__biomass_reaction_detected:
+            self.detect_biomass_reaction()
+
+        return self.__biomass_reaction
+
+    @biomass_reaction.setter
+    def biomass_reaction(self, reaction_name):
+        self.__biomass_reaction_detected = True
+        self.__biomass_reaction = reaction_name
 
     def get_flux_bounds(self, r_id):
         """ Get flux bounds for reaction
@@ -187,7 +207,7 @@ class CBModel(Model):
         if r_id in self.reactions:
             self.reactions[r_id].set_flux_bounds(lb, ub)
         else:
-            print 'No such reaction', r_id
+            raise KeyError("Reaction '{}' not found".format(r_id))
 
     def set_lower_bound(self, r_id, lb):
         """ Define lower bound for one reaction
@@ -199,7 +219,7 @@ class CBModel(Model):
         if r_id in self.reactions:
             self.reactions[r_id].lb = lb
         else:
-            print 'No such reaction', r_id
+            raise KeyError("Reaction '{}' not found".format(r_id))
 
     def set_upper_bound(self, r_id, ub):
         """ Define upper bound for one reaction
@@ -211,7 +231,7 @@ class CBModel(Model):
         if r_id in self.reactions:
             self.reactions[r_id].ub = ub
         else:
-            print 'No such reaction', r_id
+            raise KeyError("Reaction '{}' not found".format(r_id))
 
     def set_objective(self, coefficients):
         """ Define objective coefficients for a list of reactions
@@ -233,7 +253,7 @@ class CBModel(Model):
         if r_id in self.reactions:
             self.reactions[r_id].objective = coeff
         else:
-            print 'No such reaction', r_id
+            raise KeyError("Reaction '{}' not found".format(r_id))
 
     def add_reaction(self, reaction):
         """ Add a single reaction to the model.
@@ -263,19 +283,25 @@ class CBModel(Model):
             str: first reaction that matches (or else None)
         """
 
-        if not self.biomass_reaction:
+        if not self.__biomass_reaction:
             matches = [r_id for r_id, rxn in self.reactions.items() if rxn.objective]
 
             if matches:
-                self.biomass_reaction = matches[0]
-                if len(matches) == 1:
-                    print 'Biomass reaction detected:', self.biomass_reaction
-                else:
-                    print 'Multiple biomass reactions detected (first selected):', " ".join(matches)
-            else:
-                print 'No biomass reaction detected.'
+                self.__biomass_reaction = matches[0]
+                if len(matches) > 1:
+                    w ='Multiple biomass reactions detected (first selected): {}'.format(" ".join(matches))
+                    warnings.warn(w, UserWarning)
 
-        return self.biomass_reaction
+                import re
+                if not re.search("biomas|growth", self.__biomass_reaction, re.IGNORECASE):
+                    w = "Suspicious biomass reaction '{}' name".format(self.__biomass_reaction)
+                    warnings.warn(w, UserWarning)
+            else:
+                w = 'No biomass reaction detected'
+                warnings.warn(w, UserWarning)
+
+        self.__biomass_reaction_detected = True
+        return self.__biomass_reaction
 
     def add_gene(self, gene):
         """ Add a gene metabolite to the model.
@@ -307,7 +333,7 @@ class CBModel(Model):
             if gene_id in self.genes:
                 del self.genes[gene_id]
             else:
-                print 'No such gene', gene_id
+                warnings.warn("No such gene '{}'".format(gene_id), RuntimeWarning)
 
     def set_gpr_association(self, r_id, gpr):
         """ Set GPR association for a given reaction:
@@ -320,7 +346,7 @@ class CBModel(Model):
         if r_id in self.reactions:
             self.reactions[r_id].gpr = gpr
         else:
-            print 'No such reaction', r_id
+            warnings.warn("No such reaction '{}'".format(r_id), RuntimeWarning)
 
     def evaluate_gprs(self, active_genes):
         """ Boolean evaluation of the GPR associations for a given set of active genes.
@@ -339,7 +365,7 @@ class CBModel(Model):
 
         Arguments:
             r_id_num (str): id of the numerator
-            r_id_num (str): id of the denominator
+            r_id_den (str): id of the denominator
             ratio (float): ratio value
 
         Returns:
@@ -353,14 +379,14 @@ class CBModel(Model):
             self.reactions[r_id_den].stoichiometry[m_id] = -ratio
             return m_id
         else:
-            print 'Invalid reactions.'
+            raise KeyError('Invalid reactions in ratio {}/{}'.format(r_id_num, r_id_den))
 
     def remove_ratio_constraint(self, r_id_num, r_id_den):
         """ Remove a flux ratio constraint from the model.
 
         Arguments:
             r_id_num (str): id of the numerator
-            r_id_num (str): id of the denominator
+            r_id_den (str): id of the denominator
 
         """
 
@@ -369,9 +395,9 @@ class CBModel(Model):
             if m_id in self.metabolites:
                 self.remove_metabolite(m_id)
             else:
-                print 'No ratio constraint for {}/{}'.format(r_id_num, r_id_den)
+                warnings.warn('No ratio constraint for {}/{}'.format(r_id_num, r_id_den), RuntimeWarning)
         else:
-            print 'Invalid reactions.'
+            raise KeyError('Invalid reactions in ratio {}/{}'.format(r_id_num, r_id_den))
 
     def add_reaction_from_str(self, reaction_str, default_compartment=None):
         """ Parse a reaction from a string and add it to the model.
@@ -432,3 +458,233 @@ class CBModel(Model):
         """
         g_r_lookup = self.gene_to_reaction_lookup()
         return g_r_lookup[g_id]
+
+    def set_medium(self, medium):
+        """
+        Set medium for model
+        Args:
+            medium (Medium): Medium for this model
+            or raise an exception (false)
+        """
+
+        if type(medium).__name__ != "Medium":  # This is a hack for jupyter autoreloading
+            raise TypeError("Medium is not instance of <framed.model.model.Medium>")
+
+        medium_copy = medium.copy()
+        medium_copy.apply_model(self)
+
+        for r_id, bounds in medium_copy.iteritems():
+            reaction = self.reactions[r_id]
+            reaction.lb = bounds[0]
+            reaction.ub = bounds[1]
+
+
+    def get_medium(self, exchange_reaction_pattern="^R_EX_"):
+        """
+        Returns effective medium
+
+        Args:
+            exchange_reaction_pattern: Regular expression patter to search for exchange reactions
+
+        Returns:
+            Medium
+        """
+        return Medium.effective(self, exchange_reaction_pattern=exchange_reaction_pattern)
+
+class Medium(MutableMapping):
+    """
+    This class represents list of provided uptakes for the model. It inherits dictionary
+    and all the operations available in dict class
+    """
+    def __init__(self, *args, **kwargs):
+        self.store = dict()
+        self.update(dict(*args, **kwargs))
+
+    @staticmethod
+    def from_csv(path, reaction_col='reaction', lower_bound_col="lower_bound", upper_bound_col="upper_bound", sep="\t"):
+        """
+        Load medium from tab separated file
+
+        Args:
+            path: Path to medium file
+            reaction_col: Column name with reaction names
+            lower_bound_col: Column name with upper bounds
+            upper_bound_col: Column name with upper bounds
+            sep: Separator for columns
+
+        Returns: Medium
+
+        """
+        if not os.path.exists(path):
+            raise IOError(errno.ENOENT, "Media file '{}' not found".format(path), path)
+
+        medium = Medium()
+        with open(path, "r") as f:
+            header = next(f)
+            header = header.strip()
+            header = header.split("#", 1)[0]
+            header = [h.strip() for h in header.split(sep)]
+
+            for col in [reaction_col, lower_bound_col, upper_bound_col]:
+                if col not in header:
+                    raise IOError(errno.EIO, "Media file '{}' has no column '{}'".format(path, col), path)
+
+            for row in f:
+                if row.startswith("#"):
+                    continue
+
+                row = row.strip()
+                if not row:
+                    continue
+
+                row = row.split("#", 1)[0]
+                row = [c.strip() for c in row.split(sep)]
+                row = dict(zip(header, row))
+
+                medium[row[reaction_col]] = (float(row[lower_bound_col]), float(row[upper_bound_col]))
+
+        return medium
+
+
+    @staticmethod
+    def complete(model, exchange_reaction_pattern="^R_EX_"):
+        """
+        Initialize complete medium for a particular model
+
+        Arguments:
+            model (CBModel): model from which the uptake reactions are guessed
+            exchange_reaction_pattern (str): regex patter for guessing exchange reactions
+
+        Returns:
+            Medium: Complete medium for provided model
+
+        """
+        re_pattern = re.compile(exchange_reaction_pattern)
+        medium = Medium()
+        for r in model.reactions.itervalues():
+            if not re_pattern.search(r.id):
+                continue
+
+            medium[r.id] = True
+
+        return medium
+
+    @staticmethod
+    def effective(model, exchange_reaction_pattern="^R_EX_"):
+        """
+        Initialize effective medium from a provided model
+
+        Arguments:
+            model (CBModel): model from which the uptake reactions are guessed
+            exchange_reaction_pattern (str): regex patter for guessing exchange reactions
+
+        Returns:
+            Medium: Medium from provided model
+
+        """
+        re_pattern = re.compile(exchange_reaction_pattern)
+        medium = Medium()
+        for r in model.reactions.itervalues():
+            if not re_pattern.search(r.id):
+                continue
+
+            medium[r.id] = r.lb, r.ub
+
+        return medium
+
+    @staticmethod
+    def from_compounds(compounds, exchange_reaction_format="R_EX_{}_e"):
+        """
+        Initialize medium from list of compounds and a pattern for exchange reaction
+
+        Arguments:
+            compounds (list): List of compounds present in the medium
+            exchange_reaction_format (str): python format string. Use first placeholder to insert compound id
+
+        Returns:
+            Medium: Complete medium for provided model
+
+        """
+        if not iter(compounds):
+            raise TypeError("Compounds are not iterable")
+
+        medium = Medium()
+        for met in compounds:
+            r_id = exchange_reaction_format.format(met)
+            medium[r_id] = True
+
+        return medium
+
+    def __effective_bounds(self, reaction, bounds):
+        """
+        Finds effective bounds from reaction reversibility and provided bounds tuple
+
+        Arguments:
+            model (CBModel): model from which the uptake reactions are guessed
+            reaction (Reaction): Reaction for which effective bounds are to be found
+            bounds (tuple): tuple with lower and upper bounds for reaction
+
+        Returns:
+            tuple: Tuple with effective bounds for reaction
+
+        """
+        if reaction.reversible:
+            return bounds
+        else:
+            return 0.0 if bounds[0] < 0 else bounds[0], bounds[1]
+
+    def apply_model(self, model, exchange_reaction_pattern="^R_EX_"):
+        """
+        This function removes all reactions not found in the model or raises an exception
+
+        Args:
+            model (CBModel): model which is used to filter the reactions
+            reaction is not found in the model raises an exception
+        """
+        r_ids = (r_id for r_id in model.reactions if re.match(exchange_reaction_pattern, r_id))
+        for r_id in r_ids:
+            if r_id not in self:
+                self[r_id] = self.__effective_bounds(model.reactions[r_id], (0.0, 1000.0))
+            else:
+                self[r_id] = self.__effective_bounds(model.reactions[r_id], self[r_id])
+
+        r_ids = self.keys()
+        for r_id in r_ids:
+            if r_id not in model.reactions:
+                del self[r_id]
+
+    def copy(self):
+        """ Create an identical copy of the medium.
+
+        Returns:
+            Medium: medium copy
+
+        """
+
+        return deepcopy(self)
+
+    def __getitem__(self, key):
+        return self.store[self.__keytransform__(key)]
+
+    def __setitem__(self, key, value):
+        if isinstance(value, tuple) and len(value) == 2:
+            value = list(value)
+            if value[0]: value[0] = float(value[0])
+            if value[1]: value[1] = float(value[1])
+            self.store[self.__keytransform__(key)] = tuple(value)
+        elif isinstance(value, bool):
+            self.store[self.__keytransform__(key)] = (-1000.0 if value else 0.0, 1000.0)
+        else:
+            raise RuntimeError("Media value '{}' is of unknown type <{}> (only <tuple> and <bool>)".format(key, type(value)))
+
+    def __delitem__(self, key):
+        del self.store[self.__keytransform__(key)]
+
+    def __iter__(self):
+        return iter(self.store)
+
+    def __len__(self):
+        return len(self.store)
+
+    def __keytransform__(self, key):
+        return key
